requirements.txt
flask==2.3.3
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
